Move partials of base64 {en,de}code inline
Since there's only one implementation left, there is not much use in
keeping this in separate files.
